WIP move router to tRPC
